refactor: quality report モジュール分割
